Make an injector and move the flutter main content to a flutter loader
