+++ shell/platform/android/BUILD.gn
@@ -123,6 +123,7 @@ action("flutter_shell_java") {
   source_jar_path = embedding_source_jar_path
 
   sources = [
+    "io/flutter/FlutterInjector.java",
     "io/flutter/Log.java",
     "io/flutter/app/FlutterActivity.java",
     "io/flutter/app/FlutterActivityDelegate.java",
@@ -155,6 +156,10 @@ action("flutter_shell_java") {
     "io/flutter/embedding/engine/dart/DartExecutor.java",
     "io/flutter/embedding/engine/dart/DartMessenger.java",
     "io/flutter/embedding/engine/dart/PlatformMessageHandler.java",
+    "io/flutter/embedding/engine/loader/FlutterLoader.java",
+    "io/flutter/embedding/engine/loader/ResourceCleaner.java",
+    "io/flutter/embedding/engine/loader/ResourceExtractor.java",
+    "io/flutter/embedding/engine/loader/ResourcePaths.java",
     "io/flutter/embedding/engine/plugins/FlutterPlugin.java",
     "io/flutter/embedding/engine/plugins/PluginRegistry.java",
     "io/flutter/embedding/engine/plugins/activity/ActivityAware.java",
@@ -224,9 +229,6 @@ action("flutter_shell_java") {
     "io/flutter/view/FlutterNativeView.java",
     "io/flutter/view/FlutterRunArguments.java",
     "io/flutter/view/FlutterView.java",
-    "io/flutter/view/ResourceCleaner.java",
-    "io/flutter/view/ResourceExtractor.java",
-    "io/flutter/view/ResourcePaths.java",
     "io/flutter/view/TextureRegistry.java",
     "io/flutter/view/VsyncWaiter.java",
   ]
